% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fitStan}
\alias{fitStan}
\title{Results from fitting the surrogate model in Stan.}
\format{A \code{list} containing 3 variables:
\describe{
  \item{summary}{summary of the model fit for all chains merged}
  \item{c_summary}{summary for each HMC chain individually}
  \item{tm}{time taken to fit the model}
}}
\usage{
fitStan
}
\description{
Results from fitting the surrogate model in Stan.
}
\seealso{
\code{\link[rstan]{summary,stanfit-method}}
}
\keyword{datasets}
