% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPotts.R
\name{gibbsGMM}
\alias{gibbsGMM}
\title{Fit a mixture of Gaussians to the observed data.}
\usage{
gibbsGMM(y, niter = 1000, nburn = 500, priors = NULL)
}
\arguments{
\item{y}{A vector of observed pixel data.}

\item{niter}{The number of iterations of the algorithm to perform.}

\item{nburn}{The number of iterations to discard as burn-in.}

\item{priors}{A list of priors for the parameters of the model.}
}
\value{
A matrix containing MCMC samples for the parameters of the mixture model.
}
\description{
Fit a mixture of Gaussians to the observed data.
}
