% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res}
\alias{res}
\title{Simulation from the Potts model using single-site Gibbs updates.}
\format{A \code{list} containing 7 variables.}
\usage{
res
}
\description{
100 iterations of Gibbs sampling for a \eqn{500 \times 500} lattice
with \eqn{\beta=0.22} and \eqn{k=2}.
}
\seealso{
\code{\link{mcmcPotts}}
}
\keyword{datasets}
