% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resAEA}
\alias{resAEA}
\title{Results from fitting the hidden Potts model using
the approximate exchange algorithm (AEA).}
\format{A \code{list} containing 8 variables.}
\usage{
resAEA
}
\description{
Results from fitting the hidden Potts model using
the approximate exchange algorithm (AEA).
}
\seealso{
\code{\link{mcmcPotts}}
}
\keyword{datasets}
