% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simSW}
\alias{simSW}
\title{Simulation from the Potts model using Swendsen-Wang.}
\format{A \code{list} containing 3 variables:
\describe{
  \item{beta}{values of beta}
  \item{matu}{corresponding simulations of the sufficient statistic}
  \item{tm}{time taken for these simulations}
}}
\usage{
simSW
}
\description{
Simulations for a \eqn{100 \times 100} lattice for fixed values
of the inverse temperature parameter, \eqn{\beta}.
}
\seealso{
\code{\link{swNoData}}
}
\keyword{datasets}
