% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcPotts.R
\name{smcPotts}
\alias{smcPotts}
\title{Fit the hidden Potts model using approximate Bayesian computation with sequential Monte Carlo (ABC-SMC).}
\usage{
smcPotts(y, neighbors, blocks, param = list(npart = 10000, nstat = 50),
  priors = NULL)
}
\arguments{
\item{y}{A vector of observed pixel data.}

\item{neighbors}{A matrix of all neighbors in the lattice, one row per pixel.}

\item{blocks}{A list of pixel indices, dividing the lattice into independent blocks.}

\item{param}{A list of options for the ABC-SMC algorithm.}

\item{priors}{A list of priors for the parameters of the model.}
}
\value{
A matrix containing SMC samples for the parameters of the Potts model.
}
\description{
Fit the hidden Potts model using approximate Bayesian computation with sequential Monte Carlo (ABC-SMC).
}
