% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPotts.R
\name{sufficientStat}
\alias{sufficientStat}
\title{Calculate the sufficient statistic of the Potts model for the given labels.}
\usage{
sufficientStat(labels, neighbors, blocks, k)
}
\arguments{
\item{labels}{A matrix of pixel labels.}

\item{neighbors}{A matrix of all neighbors in the lattice, one row per pixel.}

\item{blocks}{A list of pixel indices, dividing the lattice into independent blocks.}

\item{k}{The number of unique labels.}
}
\value{
The sum of like neighbors.
}
\description{
Calculate the sufficient statistic of the Potts model for the given labels.
}
