% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth}
\alias{synth}
\title{Simulation from the Potts model using Swendsen-Wang.}
\format{A \code{list} containing 5 variables:
\describe{
  \item{0.22}{simulations for \eqn{\beta = 0.22}}
  \item{0.44}{simulations for \eqn{\beta = 0.44}}
  \item{0.88}{simulations for \eqn{\beta = 0.88}}
  \item{1.32}{simulations for \eqn{\beta = 1.32}}
  \item{tm}{time taken by the simulations}
}}
\usage{
synth
}
\description{
Simulations for a \eqn{500 \times 500} lattice for fixed values
of the inverse temperature parameter, \eqn{\beta}.
}
\seealso{
\code{\link{swNoData}}
}
\keyword{datasets}
