% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcPotts.R
\name{testResample}
\alias{testResample}
\title{Test the residual resampling algorithm.}
\usage{
testResample(values, weights, pseudo)
}
\arguments{
\item{values}{A vector of SMC particles.}

\item{weights}{A vector of importance weights for each particle.}

\item{pseudo}{A matrix of pseudo-data for each particle.}
}
\value{
A list containing the following elements:
\describe{
  \item{\code{beta}}{A vector of resampled particles.}
  \item{\code{wt}}{The new importance weights, after resampling.}
  \item{\code{pseudo}}{A matrix of pseudo-data for each particle.}
  \item{\code{idx}}{The indices of the parents of the resampled particles.}
  }
}
\description{
Test the residual resampling algorithm.
}

